src/requirements.txt
openai==0.27.4
discord.py==2.2.2
supervisor==4.2.5
PyNaCl==1.5.0
stability-sdk==0.6.1
Pillow==9.5.0
google-cloud-texttospeech==2.14.1
elevenlabslib==0.4.2